# be/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Untuk Login",
                "parameters": [
                    {
                        "description": "Login payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/example.ResponseLogin"
                        }
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Untuk Register",
                "parameters": [
                    {
                        "description": "Create user payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/example.ResponseRegister"
                        }
                    }
                }
            }
        },
        "/api/wallet/balance": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Untuk Mengecek sisa saldo",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/example.ResponseCheckBalance"
                        }
                    }
                }
            }
        },
        "/api/wallet/deposit": {
            "post": {
                "description": "API for terima request dari sistem lain untuk store wallet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Untuk menyetor saldo",
                "parameters": [
                    {
                        "description": "Store payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.StoreRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/example.ResponseStoreWallet"
                        }
                    }
                }
            }
        },
        "/api/wallet/withdraw": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Untuk menarik saldo",
                "parameters": [
                    {
                        "description": "Withdraw payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.WithdrawRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/example.ResponseWithdraw"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.BalanceResponse": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "number"
                },
                "last_update": {
                    "type": "string"
                },
                "wallet_id": {
                    "type": "string"
                }
            }
        },
        "dto.StoreRequest": {
            "type": "object",
            "required": [
                "amount",
                "paid_at",
                "payment_type",
                "signature_key",
                "wallet_id"
            ],
            "properties": {
                "amount": {
                    "type": "number",
                    "minimum": 1
                },
                "paid_at": {
                    "type": "string"
                },
                "payment_type": {
                    "type": "string"
                },
                "signature_key": {
                    "type": "string"
                },
                "wallet_id": {
                    "type": "string"
                }
            }
        },
        "dto.StoreResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "fee": {
                    "type": "number"
                },
                "reference_no": {
                    "type": "string"
                }
            }
        },
        "dto.TokenResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "telp": {
                    "type": "string"
                }
            }
        },
        "dto.UserCreateRequest": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password",
                "telp"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "password": {
                    "type": "string",
                    "minLength": 8
                },
                "telp": {
                    "type": "string",
                    "maxLength": 14,
                    "minLength": 8
                }
            }
        },
        "dto.UserLoginRequest": {
            "type": "object",
            "required": [
                "email_phone",
                "password"
            ],
            "properties": {
                "email_phone": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "dto.UserResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "telp": {
                    "type": "string"
                }
            }
        },
        "dto.WithdrawRequest": {
            "type": "object",
            "required": [
                "balance",
                "target_account",
                "target_bank"
            ],
            "properties": {
                "balance": {
                    "type": "number",
                    "minimum": 1
                },
                "target_account": {
                    "type": "string",
                    "maxLength": 50
                },
                "target_bank": {
                    "type": "integer"
                }
            }
        },
        "dto.WithdrawResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "fee": {
                    "type": "number"
                },
                "reference_no": {
                    "type": "string"
                },
                "target_account": {
                    "type": "string"
                },
                "target_bank": {
                    "type": "string"
                }
            }
        },
        "example.ResponseCheckBalance": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.BalanceResponse"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
        "example.ResponseLogin": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.TokenResponse"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
        "example.ResponseRegister": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.UserResponse"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
        "example.ResponseStoreWallet": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.StoreResponse"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
        "example.ResponseWithdraw": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.WithdrawResponse"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}